pyproject.toml
[tool.poetry]
name = "python-ab2-hw-airflow"
version = "0.1.0"
description = ""
authors = ["dggz <maks-laptev.03@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
apache-airflow = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
